After Shelter

+++ src/main/java/org/wecancodeit/virtualpet4/controllers/ShelterController.java
@@ -1,10 +1,12 @@
 package org.wecancodeit.virtualpet4.controllers;
 
-import org.springframework.web.bind.annotation.RestController;
+import org.wecancodeit.virtualpet4.enums.RoleEnum;
 import org.wecancodeit.virtualpet4.models.ShelterModel;
 import org.wecancodeit.virtualpet4.repositories.ShelterRepository;
+import org.wecancodeit.virtualpet4.repositories.UserRepository;
 
 import jakarta.annotation.Resource;
+import jakarta.servlet.http.HttpServletRequest;
 
 import java.util.Optional;
 
@@ -13,18 +15,22 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/shelter")
-public class ShelterController {
+public class ShelterController extends BaseController {
 
     @Resource
     private ShelterRepository shelterRepository;
+    @Resource
+    private UserRepository userRepository;
 
-    public ShelterController(ShelterRepository shelterRepository) {
+    public ShelterController(UserRepository userRepository, ShelterRepository shelterRepository) {
+        super(userRepository);
         this.shelterRepository = shelterRepository;
     }
 
     @GetMapping("")
-    public ResponseEntity<?> getAllShelters() {
+    public ResponseEntity<?> getAllShelters(HttpServletRequest request) {
         try {
+            checkAccess(RoleEnum.GUEST, request);
             return ResponseEntity.ok(shelterRepository.findAll());
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
@@ -32,20 +38,21 @@ public class ShelterController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> getShelterById(@PathVariable long id) {
+    public ResponseEntity<?> getShelterById(@PathVariable long id, HttpServletRequest request) {
         try {
+             checkAccess(RoleEnum.GUEST, request);
             return ResponseEntity.ok(shelterRepository.findById(id));
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
 
-    
     @GetMapping("/pets/{id}")
-    public ResponseEntity<?> getPetsById(@PathVariable long id) {
+    public ResponseEntity<?> getPetsById(@PathVariable long id, HttpServletRequest request) {
         try {
-              Optional<ShelterModel> opShelter = shelterRepository.findById(id);
-            if(opShelter.isPresent()){
+             checkAccess(RoleEnum.GUEST, request);
+            Optional<ShelterModel> opShelter = shelterRepository.findById(id);
+            if (opShelter.isPresent()) {
                 return ResponseEntity.ok(opShelter.get().getPets());
             }
             return ResponseEntity.ok("Shelter not found");
@@ -53,9 +60,11 @@ public class ShelterController {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
+
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteShelter(@PathVariable long id) {
+    public ResponseEntity<?> deleteShelter(@PathVariable long id, HttpServletRequest request) {
         try {
+             checkAccess(RoleEnum.ADMIN, request);
             shelterRepository.deleteById(id);
             Optional<ShelterModel> shelter = shelterRepository.findById(id);
             return ResponseEntity.ok(!shelter.isPresent());
@@ -65,11 +74,12 @@ public class ShelterController {
     }
 
     @PostMapping()
-    public ResponseEntity<?> update(@RequestBody ShelterModel shelter){
+    public ResponseEntity<?> update(@RequestBody ShelterModel shelter, HttpServletRequest request) {
         try {
+             checkAccess(RoleEnum.MANAGER, request);
             return ResponseEntity.ok(shelterRepository.save(shelter));
         } catch (Exception e) {
-             return ResponseEntity.badRequest().body(e.getMessage());
+            return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
 }
